pyproject.toml
[project]
name = "dbt-dataops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core==1.10.10",
    "dbt-postgres==1.9.0",
    "dotenv>=0.9.9",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
